requirements.txt
streamlit==1.31.0
requests==2.31.0
Pillow==10.2.0
python-docx==1.1.0
